modify time var to int32

+++ cldm/ddim_hacked.py
@@ -155,7 +155,7 @@ class DDIMSampler(object):
 
         for i, step in enumerate(iterator):
             index = total_steps - i - 1
-            ts = torch.full((b,), step, device=device, dtype=torch.long)
+            ts = torch.full((b,), step, device=device, dtype=torch.int32)
 
             if mask is not None:         # not this
                 assert x0 is not None
